Extract RTPlot's update throttle into core-ui
